Implementing the specifications on the UnitOfWork object.
This is my first pass at the real behavior that I'm trying to capture. So
right now we may find some other edge cases and such but what we've got
here is a rough start. We try and prevent duplicate changes from being
recorded as well as preventing weird situations that imply some lack
of thought on the callers side. Things like Adding something they just
Deleted and those kinds of things.