Adding Closed event to IUnitOfWork.
I debated this but it feels like the best solution to the problem. I need
to respond to the closing of the UoW in different ways (closing is
committing and rolling back by the way) For example we need to clear the
current UoW when one is closed in our provider. This avoids circular
dependencies and properly layers things.