UnitOfWorkEvents no longer decide if they apply, all of them are called.
We'll need to be able to invoke events for generic events related to
the UoW, like when it's been started. And the model of doing the whole
do we apply thing just won't work there (we don't have instances yet and
we only need to get called once) So it's either have two separate event
type models or one and delegate if we need finer grained control.

I chose the latter approach. It makes sense because we'll only need a pair
of these in an application at *most*, anymore and I think there's something
strange going on. In fact, most people would probably need to create their
own hybrid implementation, rather than relying on two.