Fixing things when inside a TransactionScope.
Apparently closing the session is bed because NH enlists in the scope and
tries to "do stuff" with the session when it's completed, that fails
because we're closed. So we don't close anymore.

I'm wondering if we should enlist as well to make it easier on people who
don't want to deal with TransactionScope's in their implementations of
the IUnitOfWorkEvents interface.