Making things integrate with System.Transactions better.
I setup an EnlistmentNotifications instance if we're inside a scope and
manage the UnitOfWork off of that. One thing we had to do was skip
the calls for NH since NH will hook into the same scope and we don't want
to do everything twice.

We also don't Rollback in Dispose like we did earlier. We leave that to
the Scope to decide.

This may be the completely wrong behavior. I'm just trying to setup a
convention here so feel fry to cry foul if you see something wrong. It
might be a good idea to also prevent the user from invoking the IUnitOfWork
Commit and Rollback directly in such cases.