Making NH play nicely with ambient transactions.
We don't start an NH transaction if we're in a distributed one already
and we also Flush in the Rollback (and therefore the Dispose) if that's
also the case.

Not too stoked about how I did this mostly because I'm a little sketchy
on how some of the System.Transactions stuff is working. Like why does
the thread pool get involved and stuff and why we run into strange errors
w/o the Flush. I think in the future it may be a good idea to have a
separate Dispose on the Events as well as Rollback just so the two can be
handled separately if necessary.