Ambient transaction scoping and provider.
We need the ability to behave smartly under an ambient transaction. For
example we need to separate the UoW if we'd like so that we can rollback
the database transaction and still commit our scope. Also it'd be nice to
behave nicely when we receive a batch of a specific message and our UoW
is tied to the outer scope (this way we can effectively re-use the same
session and connection, etc...)